requirements.txt
Django==5.0.1
django-extensions==3.2.3
pygraphviz==1.12
django-debug-toolbar==4.2.0
django-tinymce>=3.0.0
Pillow==8.4.0   
